# docs/swagger.yaml
definitions:
  dto.DollarResponse-dto_Dollar:
    properties:
      code:
        type: integer
      data:
        properties:
          buy:
            $ref: '#/definitions/dto.Price'
          name:
            type: string
          sell:
            $ref: '#/definitions/dto.Price'
          timestamp:
            type: string
        type: object
      message:
        type: string
    type: object
  dto.DollarResponse-dto_Page:
    properties:
      code:
        type: integer
      data:
        properties:
          name:
            type: string
          route:
            type: string
        type: object
      message:
        type: string
    type: object
  dto.Price:
    properties:
      value:
        type: number
      valueText:
        type: string
    type: object
info:
  contact:
    email: brendasanchez9310@gmail.com
    name: dusk
  description: cotizaciones del dolar
  title: ws-money
paths:
  /v1/dollar:
    get:
      consumes:
      - application/json
      description: get dollars by param web
      parameters:
      - description: pages to ws
        example: DOLAR_HOY
        in: query
        name: web
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DollarResponse-dto_Dollar'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Consigue los tipos de cambio del dolar
      tags:
      - dollar
  /v1/pages:
    get:
      consumes:
      - application/json
      description: get pages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DollarResponse-dto_Page'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Consigue las paginas donde traera los precios
      tags:
      - dollar
swagger: "2.0"